Revert "refazendo rotas"
